LayoutTests/editing/find/cocoa/find-in-page-in-hidden-until-found.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/dump-as-markup.js"></script>
<script src="../../../resources/ui-helper.js"></script>
</head>
<body>
    <div id="hiddenUntilFound" hidden="until-found">
        <div id="editor" contenteditable>
            <p>Apple banana orange.</p>
        </div>
    </div>
    <div id="otherHiddenUntilFound" hidden="until-found">
        <div id="nestedHiddenUntilFound" hidden="until-found">
            <div id="otherEditor" contenteditable>
                <p>Kiwi banana pear.</p>
            </div>
        </div>
    </div>
    <div id="log"></div>
</body>
<script>
Markup.waitUntilDone();
Markup.description("Verifies that find and replace can be used to replace words in an editable area. This test requires WebKitTestRunner.");

function waitForReveal(element) {
    return new Promise((resolve, reject) => {
        element.onbeforematch = () => {
            if (!element.hidden)
                reject();
            setTimeout(() => {
                if (!element.hidden)
                    resolve();
            }, 0);
        }
    });
}

onload = async () => {
    testRunner.findStringMatchesInPage("orange", []);
    testRunner.indicateFindMatch(0);
    await Promise.all([waitForReveal(hiddenUntilFound), UIHelper.ensurePresentationUpdate()]);
    testRunner.replaceFindMatchesAtIndices([0], "apricot", false);
    Markup.dump("editor", "After replacing 'orange' with 'apricot'");

    if (hiddenUntilFound.hidden) {
        log.textContent += "FAIL because first element didn't expand after first replacement.\n";
    }

    if (!otherHiddenUntilFound.hidden || !nestedHiddenUntilFound.hidden) {
        log.textContent += "FAIL because second element expanded after first replacement, despite word not matching in second element.\n";
    }

    hiddenUntilFound.hidden = "until-found";

    testRunner.findStringMatchesInPage("banana", []);
    testRunner.indicateFindMatch(0);
    await Promise.all([waitForReveal(hiddenUntilFound), UIHelper.ensurePresentationUpdate()]);
    testRunner.indicateFindMatch(1);
    await Promise.all([waitForReveal(otherHiddenUntilFound), waitForReveal(nestedHiddenUntilFound), UIHelper.ensurePresentationUpdate()]);
    testRunner.replaceFindMatchesAtIndices([0, 1], "watermelon", false);
    Markup.dump("editor", "First editor after replacing 'banana' with 'watermelon'");
    Markup.dump("otherEditor", "Second editor after replacing 'banana' with 'watermelon'");

    if (hiddenUntilFound.hidden) {
        log.textContent += "FAIL because first element didn't expand after second replacement.\n";
    }

    if (otherHiddenUntilFound.hidden || nestedHiddenUntilFound.hidden) {
        log.textContent += "FAIL because second element didn't expand after second replacement.\n";
    }

    if (log.textContent)
        Markup.dump("log");

    Markup.notifyDone();
};
</script>
</html>
